requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
openai
faiss-cpu>=1.12.0
fastapi
uvicorn[standard]
pydantic
httpx
requests
python-dotenv
pypdf
tqdm
numpy
tiktoken